#183 <Feat> Рефакторинг
close #183
