+++ package.json
@@ -40,6 +40,7 @@
         "eslint-plugin-react": "^7.32.2",
         "eslint-plugin-standard": "^5.0.0",
         "markdown-it": "^13.0.1",
+        "markdown-it-link-attributes": "^4.0.1",
         "prismjs": "^1.29.0",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
