+++ src/api/moderation.ts
@@ -1,13 +1,10 @@
 import { GPTModeration } from "$/entity/GPT";
 
-export function moderationText(
-  input: string,
-  apiKey: string
-): Promise<GPTModeration> {
+export function moderationText(input: string): Promise<GPTModeration> {
   return fetch("https://api.openai.com/v1/moderations", {
     method: "POST",
     headers: {
-      Authorization: "Bearer " + apiKey,
+      Authorization: "Bearer ",
       "Content-Type": "application/json",
     },
     body: JSON.stringify({ input, model: "text-moderation-latest" }),
