+++ src/components/Messenger/MessengerList/Message/Message.tsx
@@ -3,7 +3,7 @@ import React, { memo } from "react";
 import { classNames, Div, IconButton, Text } from "@vkontakte/vkui";
 import { Icon24CheckCircleOutline } from "@vkontakte/icons";
 
-import { ChatGpt, GptMessage } from "$/entity/GPT";
+import { GptMessage } from "$/entity/GPT";
 
 import { vkUser } from "$/entity/user";
 
@@ -14,9 +14,10 @@ import { WarningTooltip } from "./WarningTooltip";
 
 import classes from "./Message.module.css";
 import { Copy } from "$/components/Copy";
+import { ChatGptTemplate } from "$/entity/GPT/ChatGptTemplate";
 
 interface IProps {
-  chatGpt: ChatGpt;
+  chatGpt: ChatGptTemplate;
   isDisabled: boolean;
   message: GptMessage;
 }
