+++ src/entity/GPT/ChatGptFree.ts
@@ -1,0 +1,3 @@
+import { ChatGptTemplate } from "$/entity/GPT/ChatGptTemplate";
+
+export class ChatGptFree extends ChatGptTemplate {}
