+++ src/entity/GPT/ChatGptLesson.ts
@@ -1,0 +1,3 @@
+import { ChatGptTemplate } from "$/entity/GPT/ChatGptTemplate";
+
+export class ChatGptLesson extends ChatGptTemplate {}
