+++ src/entity/messages/types.ts
@@ -2,7 +2,7 @@ export type CreateMessageParams = {
   historyId: string;
   content: string;
   role: string;
-  isError: boolean;
+  error: boolean;
   isFailedModeration: boolean;
   lastUpdated: Date;
 };
@@ -11,6 +11,6 @@ export type RemoteMessage = {
   id: string;
   content: string;
   role: string;
-  isError: boolean;
+  error: boolean;
   isFailedModeration: boolean;
 };
