+++ src/panels/Home/FreeDialogBlock/FreeDialogBlock.tsx
@@ -7,8 +7,6 @@ import { ChatGPTLogo } from "$/icons";
 import { CardBlock } from "$/components/CardBlock";
 
 import classes from "./FreeDialogBlock.module.css";
-import { lessonsController } from "$/entity/lessons";
-import { chatGpt } from "$/entity/GPT";
 
 interface IProps {
   goToFreeDialog: () => void;
@@ -26,19 +24,7 @@ function FreeDialogBlock({ goToFreeDialog }: IProps) {
             mode="outline"
             size="m"
             after={<Icon24ArrowRightSquareOutline />}
-            onClick={() => {
-              const currentChapter = lessonsController.currentChapter.get();
-              const currentLesson = lessonsController.currentLesson.get();
-              if (currentChapter || currentLesson) {
-                lessonsController.clearLesson();
-                lessonsController.clearChapter();
-              }
-
-              chatGpt.clearMessages();
-              chatGpt.abortSend();
-
-              goToFreeDialog();
-            }}
+            onClick={goToFreeDialog}
           >
             Начать диалог
           </Button>
